requirements.txt
pandas>=2.2.0
numpy>=1.26.0
Faker>=24.0.0
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
requests>=2.32.3
PyYAML>=6.0.2
# Testing (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# Airflow opsional (instal terpisah karena dependency besar)
# apache-airflow==2.10.2